refactor: extract video component
